+++ src/struct_test.h
@@ -10,7 +10,7 @@ using namespace Rcpp;
 
 double struct_test_mean(NumericVector & y, std::string funname, bool self=false);
 double struct_test_var(NumericVector & y, std::string funname, bool self=false);
-NumericVector ego_variance(const arma::sp_mat & G, const NumericVector & Y,
+NumericVector ego_variance(const arma::sp_mat & graph, const NumericVector & Y,
                            std::string funname);
 NumericVector hatf(const arma::sp_mat & G, const  NumericVector & Y,
                    std::string funname);